requirements.txt
pytesseract
Pillow
evdev